created new connection file
The main connection code has been switched to a new file. All commands
are working except or quit…
